requirements.txt
streamlit>=1.28.0
langgraph>=0.0.20
langchain-groq>=0.0.1
playwright>=1.40.0
chromadb>=0.4.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0